Fixed Float in visits/models

+++ visits/models.py
@@ -16,7 +16,7 @@ class Visit(models.Model):
         'maintenance.Contract', null=True, on_delete=models.SET_NULL)
     visit_date = models.DateField(default=date.today)
     crewsize = models.CharField(max_length=128, blank=False)
-    total_man_hours = models.IntegerField(blank=False)
+    total_man_hours = models.FloatField(blank=False)
     notes = models.CharField(max_length=1024, null=True, blank=True)
     date_created = models.DateTimeField(auto_now_add=True, null=True)
